Change Shape.tag to Int64

+++ Patchworks/Models/Shape+Convenience.swift
@@ -10,7 +10,7 @@ import UIKit
 import CoreData
 
 extension Shape {
-    @discardableResult convenience init(context: NSManagedObjectContext = CoreDataStack.context, tag: Int16, image: Data) {
+    @discardableResult convenience init(context: NSManagedObjectContext = CoreDataStack.context, tag: Int64, image: Data) {
         self.init(context: context)
         self.tag = tag
         self.image = image
